backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pandas>=1.5.0
numpy>=1.24.0
scikit-learn
plotly
textblob
transformers
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
databricks-sdk
pytest
httpx
